Updated project struct
