+++ configs/application.yaml
@@ -1,0 +1,5 @@
+name: InfiniteUGCTool
+description: An interactive CLI to bookmark, clone and manage Halo Infinite UGC files
+version: 0.5.0
+author: Alexis Bize
+repository: https://github.com/Alexis-Bize/InfiniteUGCTool
