--- internal/shared/libs/halowaypoint/main.go
@@ -1,1 +1,0 @@
-package halowaypoint
