--- internal/shared/libs/msa/config.go
@@ -1,1 +1,0 @@
-package msa
