--- internal/shared/modules/helpers/release/release.types.go
@@ -1,7 +1,0 @@
-package release
-
-type Release struct {
-	TagName string `json:"tag_name"`
-}
-
-type Releases []Release
